requirements.txt
Flask
requests
python-dotenv
flask-swagger-ui
